Test - Make compile of jars optional

+++ script/multi-machine/setup.py
@@ -7,15 +7,16 @@ from machine_info import MASTER_IP_ADDRESS, MASTER_PORT, WORKER_IP_ADDRESS, WORK
 TESTCASE_DIRECTORY = 'script/multi-machine/testcase'
 SBT_BIN = '/home/cyan/.sbt.bin'
 
-def setup_master():
+def setup_master(compile_jar=True):
     # Compile the project on the master machine because there is no java on the ssh bridge machine.
     ssh = createSSHClient(MASTER_IP_ADDRESS, MASTER_PORT)
-    exec_command_blocking(ssh, 'rm -f worker.jar master.jar')
     exec_command_blocking(ssh, 'rm -rf 332project && git clone https://github.com/seonghyun26/332project.git')
-    exec_command_blocking(ssh, f'cd /home/cyan/332project && {SBT_BIN}/sbt "master/assembly"')
-    exec_command_blocking(ssh, f'cd /home/cyan/332project && {SBT_BIN}/sbt "worker/assembly"')
-    exec_command_blocking(ssh, f'mv /home/cyan/332project/master/target/scala-2.12/master.jar /home/cyan/master.jar')
-    exec_command_blocking(ssh, f'mv /home/cyan/332project/worker/target/scala-2.12/worker.jar /home/cyan/worker.jar')
+    if compile_jar == True:
+        exec_command_blocking(ssh, 'rm -f worker.jar master.jar')
+        exec_command_blocking(ssh, f'cd /home/cyan/332project && {SBT_BIN}/sbt "master/assembly"')
+        exec_command_blocking(ssh, f'cd /home/cyan/332project && {SBT_BIN}/sbt "worker/assembly"')
+        exec_command_blocking(ssh, f'mv /home/cyan/332project/master/target/scala-2.12/master.jar /home/cyan/master.jar')
+        exec_command_blocking(ssh, f'mv /home/cyan/332project/worker/target/scala-2.12/worker.jar /home/cyan/worker.jar')
 
 def setup_workers():
     client = createSSHClient(MASTER_IP_ADDRESS, MASTER_PORT)
@@ -33,8 +34,8 @@ def setup_workers():
                 putFile(client, f'{TESTCASE_DIRECTORY}/{config_file_name}', f'/home/cyan/testcase/{config_file_name}')
             client.close()
 
-def setup_machines():
-    setup_master()
+def setup_machines(compile_jar=True):
+    setup_master(compile_jar=compile_jar)
     setup_workers()
 
 if __name__ == '__main__':
